% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portOptimizer.r
\name{addTermToObjective}
\alias{addTermToObjective}
\title{Create / add a term to an objective definition}
\usage{
addTermToObjective(.lst, name = "", type = "constraint", funName,
  params = list(x), min = -Inf, max = Inf)
}
\arguments{
\item{.lst}{If missing, this is ignored. If present, the function appends a term (a list) to this.}

\item{name}{A description of the term. Useful for provide a breakdown of each term's contribution to the objective function.}

\item{type}{Either 'maximize', 'minimize', or 'contraint'. If a constraint, then the `min` and/or `max` should come into play.}

\item{funName}{A function}

\item{params}{The (not x) parameters needed by `funName`.  All functions will need the weights (x). We assume that's the first parameter and it should be ommitted from `params``}

\item{min}{Minimum allowed value for constraints. Default is -Inf.  Ignored if -Inf.}

\item{max}{Maximum allowed value for constraints. Default is Inf. Ignored if Inf.}
}
\value{
List
}
\description{
Create / add a term to an objective definition
}
\examples{
addTermToObjective(name = "MaxReturn", type="maximize", funName=calc_port_return, params = list(expected.returns))
addTermToObjective(name = "Risk", type = "constraint", funName = calc_port_risk, params = list(as.matrix(expected.cov)), min = -Inf, max = calc_port_risk(benchX, as.matrix(expected.cov)))
}
