% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portOptimizer.r
\name{calc_constraint}
\alias{calc_constraint}
\title{Calculate penalty associated with a constraint in an object definition}
\usage{
calc_constraint(obj, x)
}
\arguments{
\item{obj}{List defining constraint}

\item{x}{vector of portfolio weights}
}
\value{
scalar
}
\description{
Only returns squared error. The penaltyFactor has not been applied.
}
\examples{
calc_constraint(obj, x)
}
