% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portOptimizer.r
\name{calc_maxmin}
\alias{calc_maxmin}
\title{Calculate objective max or min associated with an object definition}
\usage{
calc_maxmin(obj, x)
}
\arguments{
\item{obj}{List defining constraint}

\item{x}{vector of portfolio weights}
}
\value{
scalar
}
\description{
Calculate objective max or min associated with an object definition
}
\examples{
calc_maxmin(obj, x)
}
