% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_min_nonzero_weight}
\alias{calc_min_nonzero_weight}
\title{Minimum nonzero portfolio weight
Function works for long-only portfolios and would need to be modified for long-short.
If LB is missing, the minimum nonzero weight is returned.  If LB is not missing, then
the smallest nonzero weights for asset classes without a zero LB is returned.}
\usage{
calc_min_nonzero_weight(x, LB)
}
\arguments{
\item{x}{Weight vector of the portfolio}

\item{LB}{Optional lower bound of weights}
}
\value{
scalar
}
\description{
Minimum nonzero portfolio weight
Function works for long-only portfolios and would need to be modified for long-short.
If LB is missing, the minimum nonzero weight is returned.  If LB is not missing, then
the smallest nonzero weights for asset classes without a zero LB is returned.
}
\examples{
calc_min_nonzero_weight(x)
}
