% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_nnonzeroassets}
\alias{calc_nnonzeroassets}
\title{Number of nonzero-weighted asset}
\usage{
calc_nnonzeroassets(x)
}
\arguments{
\item{x}{Weight vector of the portfolio}
}
\value{
scalar
}
\description{
Number of nonzero-weighted asset
}
\examples{
calc_nnonzeroassets(x)
}
