% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portOptimizer.r
\name{calc_objective}
\alias{calc_objective}
\title{Calculate an objective given a objective definition}
\usage{
calc_objective(x, obj_definition, penaltyFactor = 25000000000)
}
\arguments{
\item{x}{Vector of portfolio weights}

\item{obj_definition}{A list serving as an object definition.}

\item{penaltyFactor}{A large value by which squared errors are multiplied to produce a penalty}
}
\value{
value
}
\description{
Calculate an objective given a objective definition
}
\examples{
calc_objective(x, obj_definition)
}
