% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_port_return}
\alias{calc_port_return}
\title{Portfolio Return}
\usage{
calc_port_return(x, expectedReturns)
}
\arguments{
\item{x}{Weight vector of the portfolio}

\item{expectedReturns}{Vector of the expected returns}
}
\value{
scalar
}
\description{
Portfolio Return
}
\examples{
calc_port_return(x, expectedReturns)
}
