% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_port_risk}
\alias{calc_port_risk}
\title{Portfolio Risk (Standard Deviation)}
\usage{
calc_port_risk(x, expectedCov)
}
\arguments{
\item{x}{Weight vector of the portfolio}

\item{expectedCov}{Covariance matrix with dim(length(x), length(x))}
}
\value{
scalar
}
\description{
Portfolio Risk (Standard Deviation)
}
\examples{
calc_port_risk(x, expectedCov)
}
