% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_port_TE}
\alias{calc_port_TE}
\title{Tracking error}
\usage{
calc_port_TE(x, benchX, expectedCov)
}
\arguments{
\item{x}{Weight vector of the portfolio}

\item{benchX}{Weight vector of the benchmark}

\item{expectedCov}{Covariance matrix with dim(length(x), length(x))}
}
\value{
scalar
}
\description{
Tracking error
}
\examples{
calc_port_TE(x, benchX, expectedCov)
}
