% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_turnover}
\alias{calc_turnover}
\title{Turnover
Turnover is defined such that if the entire portfolio is sold and repurchases, the value is 1.}
\usage{
calc_turnover(x, curWeights)
}
\arguments{
\item{x}{Weight vector of the proposed portfolio}

\item{curWeights}{Weight vector of the current portfolio}
}
\value{
a scalar
}
\description{
Turnover
Turnover is defined such that if the entire portfolio is sold and repurchases, the value is 1.
}
\examples{
calc_turnover(x, curWeights)
}
