% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioFunctions.r
\name{calc_weighted}
\alias{calc_weighted}
\title{Weighted Average}
\usage{
calc_weighted(x, weights)
}
\arguments{
\item{x}{Weight vector of the proposed portfolio}

\item{weights}{Coefficients to use as weights}
}
\value{
Scalar
}
\description{
A simple function with many uses.  Besides the obvious use of weighted average to calculate something like portfolio yield,
this can be used to find the weight of a group (set) of assets. For example, if the weights vector has 1 for each asset in
a group and zero for the others, then the result will be the allocation to that group.  Consider a set of many assets
including US Large and US Small.  If the US Large asset has a weight of 0.25 and the US Small has a weight of -0.25, then
as long as the result is >= 0, US Large will be at least 75% of the sum of US Large plus US Small. Such an approach is
useful for setting a variety of constraints.  To arrive at those coefficients we used basic algebra:
     USL / (USL + USS) >= 0.75  (where USL is US Large and USS is US Small)
     USL >= 0.75USL + 0.75USS so
     0.25USL - 0.75USS >= 0.
so for this constraint we create a vector of coefficients (weights) with +0.25 for USL, -0.75 for USS, and 0 for the others.
}
\examples{
calc_weighted(x, weights)
}
