% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portOptimizer.r
\name{maximize_objective}
\alias{maximize_objective}
\title{Portfolio optimization using a pairwise algorithm}
\usage{
maximize_objective(x, LB, UB, objFun, ...)
}
\arguments{
\item{x}{Initial set of weights. Must sum to 1 (or whatever sum is desired).  Weights across accounts must be correct.}

\item{LB}{Box minimum}

\item{UB}{Box maximum}

\item{objFun}{Objective function. First parameter should be the portfolio weights}

\item{...}{Parameters to be passed to the objective function}

\item{maxdelta}{Initial setting of delta.}
}
\value{
List with MaxObj the value of the objective; x the corresponding portfolio weights, and Iter the number of iterations
}
\description{
This will attempt to maximize a supplied objective function within box constraints.
The algorithm begins by finding all the asset classes with weights that can be increased and those
that can be decreased.  Using pairs of classes, it increases one and decreases the other
finding the best pair to change.  If a pair improves the objective value, the process is repeated.  If no pair is found
the maxdelta is halved and the process is repeated until the delta is <= 0.00001.
}
\examples{
maximize_pairwise(x, LB, UB, objFun, maxdelta, ...)
}
